requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
numpy==1.21.6
pandas==1.1.5
python-dateutil==2.8.2
pytz==2021.3
requests==2.28.0
six==1.16.0
urllib3==1.26.9
Flask==2.2.2
Flask-Mail==0.9.1
